Better handling of nodedown errors.
The fabric controllers now make use of
rexi_monitor to receive `DOWN` error messages
and attempt to continue if possible by removing
the dead shards.

So for example the W constant is now used by
fabric_db_create. If the quorum is reached an HTTP
201 is returned, if not an HTTP 201. The calls will
no longer hang when a node is unavailable.

For some, such as the views with implicit R=1 it won't
much matter, though it might help  calls to fail faster.

BugzID:12220