Support for read your writes consistency
On updates, an encoding of the shards that participate, along with the current
sequence, is returned to the client in a header X-Couch-RYW. If this header is then
passed in to subsequent view queries, the encoded shards will be used.
Multiple X-Couch-RYW headers can be used.

This encoded list of {shard, seq} is also returned on view requests, even if an
X-Couch_RYW header was not specified from a previous update. In the case
stale=ok is not used, and view groups are kicked off these sequences will provide
something between stale=ok and the current sequence on subsequent calls.

Multiple X-Couch_RYW headers can also be specified, in which case the maximum sequence
number is used for shards that are in the intersection of the two sets.

BugzID: 11393