Add open Db to filter call
The custom filter fun had a Db included that was created by configure_filter when called by
chttpd:handle_changes_req. It was needed for a fabric clustered call to retrieve the DDoc.
This patch removed that Db from the custom filter spec and adds the shard local open db to the
filter call.

BugzID: 15067