Filtered continuous changes feeds now report progress when filters fail
At times, some filters are very strict and thousands of docs might
pass thru without success. If a timeout occurs the changes feed could
be requested again starting from the same sequence. This patch fixes this
by reporting a seq back to the controller even when the filter fails. The tricky
part is that this subverts the normal rexi mechanism of using inactivity
to determine a timeout has occured. To solve this the time remaining until the
timeout is tracked in the handle message function and returned to rexi to
use as the new per message timeout. This effectively allows rexi to ignore these
messages for timeout purposes.

BugzID: 17709