Filtered continuous changes feeds report progress when filters fail
At times some filters are very strict and thousands of docs might
pass thru without success. If a timeout occurs the changes feed could
be requested again starting from the same sequence. This patch fixes this
by reporting a seq back to the controller even when the filter fails. The tricky
part is that this subverts the normal rexi mechanism of using inactivity
to determine a timeout has occured for purposes of a heartbeat. To solve this
an additional "time since last real change" is tracked in the controller.

BugzID: 17709