Send a no_pass message to fabric controller
When a filter function fails for a given document send the sequence back to
the controller. This enables long running changes feeds that may timeout
to return the last sequence processed to the caller. A count is maintained in
order to only send a message every so many times.

BugzID: 17709