Change write quorum error logs into metrics
The global_changes work uncovered the fact that the common "write quourm
failed" log messages happen for run of the mill conflict errors. Rather
than try and change the logic around to fix the error this just adds
three metrics to cover the events so that we can monitor it in the
future.

BugzId: 23977