Always send detailed sequence information
Previously we were only sending the full triplet of sequence number,
UUID, and epoch node when transmitting an actual row to the coordinator.
The 'complete' message at the end of the feed omitted the epoch owner,
and the 'no_pass' message just sent the sequence number.  The 'complete'
messages are used in continuous _changes feeds as well, so the end
result is that we had pretty spotty coverage for epoch node information
in the sequences.

BugzID: 27193