Indicate when read quorums are met
Currently there is no indication if a read operation has met the
requested quorum level. This patch and a corresponding patch to dbcore
enable this indication by adding a special document member _r_met that
is set to false when quorum is not met. When the requested quorum is met
the field does not exist. This is to ensure no changes for normal use
cases.

Unfotunately this change breaks replication to any node or cluster that
doesn't have these patches. I don't really expect this to be merged as
is but the tests were so pretty I couldn't help but finish it off.

BugzId: 26369