Allow for read repair localized to request
Our default behavior is to commit the repair back to the shard files. In
deployments with highly concurrent reads this can have a material impact
on the total write throughput for the cluster. An alternative is to use
the repair logic in the coordinator but not commit the final resolution
back to disk, and instead wait for internal replication to fix up the
inconsistent shard copy. This patch adds a config flag to enable that
option:

[fabric]
commit_read_repair = false

BugzID: 33171